Combine files into a view model
